# whatsapp-clone/src/main/java/com/xcoder/whatsapp/clone/common/StringResponse.java
package com.xcoder.whatsapp.clone.common;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class StringResponse {
    private String response;
}